logger => withLogger
