+++ README.md
@@ -24,7 +24,7 @@ npm i <a href="https://www.npmjs.com/package/@hyperapp/logger">@hyperapp/logger<
 Then with a module bundler like [rollup](https://github.com/rollup/rollup) or [webpack](https://github.com/webpack/webpack) use as you would anything else.
 
 ```jsx
-import logger from "@hyperapp/logger"
+import withLogger from "@hyperapp/logger"
 ```
 
 ### Browser
@@ -35,12 +35,12 @@ Download the minified library from the [CDN](https://unpkg.com/@hyperapp/logger)
 <script src="https://unpkg.com/@hyperapp/logger"></script>
 ```
 
-You can find the library in `window.logger`.
+You can find the library in `window.withLogger`.
 
 ## Usage
 
 ```js
-logger(options)(app)(state, actions, view, document.body)
+withLogger(options)(app)(state, actions, view, document.body)
 ```
 
 ### Options
@@ -50,7 +50,7 @@ logger(options)(app)(state, actions, view, document.body)
 Use it to customize the log function.
 
 ```js
-logger({
+withLogger({
   log(prevState, action, nextState) {
     // format and send your log messages anywhere you like
   }
