+++ test/logger.test.js
@@ -1,5 +1,5 @@
 import { h, app } from "hyperapp"
-import logger from "../src"
+import withLogger from "../src"
 
 const defaultConsole = console
 
@@ -8,7 +8,7 @@ afterEach(() => {
 })
 
 test("without actions", done =>
-  logger()(app)(undefined, undefined, () => done()))
+  withLogger()(app)(undefined, undefined, () => done()))
 
 test("log", done => {
   console = {
@@ -19,7 +19,7 @@ test("log", done => {
     }
   }
 
-  logger()(app)(
+  withLogger()(app)(
     {},
     {
       foo: () => state => state
@@ -28,7 +28,7 @@ test("log", done => {
 })
 
 test("custom log function", done =>
-  logger({
+  withLogger({
     log(state, action, nextState) {
       expect(state).toEqual({ value: 0 })
       expect(action).toEqual({ name: "inc", data: 2 })
@@ -45,7 +45,7 @@ test("custom log function", done =>
   ).inc(2))
 
 test("state slices", done => {
-  const actions = logger({
+  const actions = withLogger({
     log(state, action, nextState) {
       switch (action.name) {
         case "hello":
@@ -80,7 +80,7 @@ test("state slices", done => {
 })
 
 test("doesn't interfere with state updates", done => {
-  const actions = logger({
+  const actions = withLogger({
     log(state, action, nextState) {}
   })(app)(
     {
@@ -117,7 +117,7 @@ test("doesn't interfere with state updates", done => {
 
 test("doesn't interfere with custom container", done => {
   document.body.innerHTML = "<main></main>"
-  logger({
+  withLogger({
     log(state, action, nextState) {}
   })(app)(
     {},
